Uploading the sample app
Thank you for attending the interview at Xceed Ventures. Its seems that
you have pass the interview on User Experience Software Engineer or you
explicity asked to see an example of or Uxer Experience code. Please
follow the instructions below for your code assignment:
Code Assignment:
1) Review the application in this reprosotiry
https://github.com/memonono/SampleXceedUX.
2) To run the application, you need a web server or and IDE that have
web server. It is built using Backbone.Js and Marionette.Js
3) There is a folder called screenshots that contains the original
screen designs so you can refer to it.
4) The application contains multiple Timelines and the josons are
provided to you in a folder, please use the JSONs to show the timelines
in the application.
5) The menues on the right side of the screen, are hardcoded. Please
load the main menu from a JSON instead.
6) Implment the search functionality by filtering the timeline. So if I
click on the search tab, I can search within the timeline text.
7) Create a GitHub account and upload your work.
8) Send me an email summarizing what you did and upload the GitHub link.
Please don’t attach code in the email.
9) You will be evaluated on how fast you can submit and how good quality
is your code.